Allow ARC to handle scatter metadata
Some metadata types are not required or not easily to use scatter ABD. So to
allow ARC to accommodate both types of metadata, we introduce a new flag
ARC_BUFC_SCATTER to indicate the type of the ABD. This flag only applies to
metadata because userdata always use scatter ABD.

Now that we cannot use `type == ARC_BUFC_METADATA' to check whether is
metadata, we need to use the macro ARC_BUFC_IS_META.

For things that doesn't care about the scatter flag. For them, we use the macro
ARC_BUFC_TYPE_MASK to mask out the scatter flag. We also create two helper
macros ARCS_LIST and ARCS_LSIZE to do the masking for us.

Signed-off-by: Chunwei Chen <tuxoko@gmail.com>